requirements.txt
alembic==1.6.2
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.0
Jinja2==3.0.0
limits==1.5.1
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.12.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0